refactor: list view tab 이벤트 등록 로직 list-view.js로 옮김
